requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
future==0.18.2
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
psycopg2-binary
Gunicorn
